refactor(ui): Extract plugin utilities
These include utility functions related to validating required plugin
options, getting the default values for plugins, and creating plugin
payloads.
